refactor: :recycle: changes to support the next feature
the next feature here is the brute-force recovering when
some share pieces are missing. in this commit we also add
the `verify` operation for secret recovering verification.
shareholders can verify that indeed the secret was recovered
with success, cause it encrypts their shares -- with
H(H(secret)) -- and authenticates/signs them as well -- with
H(H(H(secret))).

Signed-off-by: Marco Aurélio da Silva <marcoonroad@gmail.com>